move error to separate folder
